requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
GDAL==3.11
